requirements.txt
fastapi
uvicorn
pydantic
google-auth
python-jose
sqlalchemy[asyncio]
asyncpg
firebase-admin
python-dotenv
passlib
bcrypt
google-cloud-secret-manager
requests
playwright
email-validator
pytest-asyncio
redis[asyncio]